WIP: feat: basic support for multiple repositories
`repositoryName` inside of `slicemachine.config.json` can now either be
a string or an array of strings. We could've of course just changed the
type to only support an array of strings, but IMO a few
`typeof === string` in the codebase is less painful than literally
having to update every config file ever made.